requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
google-genai>=1.0
python-dotenv>=1.0

pytest>=8.0
pytest-cov>=5.0
httpx>=0.27
